requirements.txt
Django==6.0.2
djangorestframework==3.14.0
django-filter==23.2
Pillow==10.0.0
python-dotenv==1.0.0
